chore(Home): refactor home page to separate the file into components
