+++ public/rss/feed.xml
@@ -4,7 +4,7 @@
         <title>Spencer Sharp</title>
         <link>undefined</link>
         <description>Your blog description</description>
-        <lastBuildDate>Tue, 27 Jun 2023 09:24:59 GMT</lastBuildDate>
+        <lastBuildDate>Tue, 27 Jun 2023 11:16:59 GMT</lastBuildDate>
         <docs>https://validator.w3.org/feed/docs/rss2.html</docs>
         <generator>https://github.com/jpmonette/feed</generator>
         <image>
